+++ src/evals/metrics/mia/min_k_plus_plus.py
@@ -30,9 +30,9 @@ class MinKPlusPlusAttack(MinKProbAttack):
 
         # Handle numerical stability
         sigma = torch.clamp(sigma, min=1e-6)
-        scores = (target_prob.cpu().numpy() - mu.cpu().numpy()) / torch.sqrt(
-            sigma
-        ).cpu().numpy()
+        scores = (
+            target_prob.float().cpu().numpy() - mu.float().cpu().numpy()
+        ) / torch.sqrt(sigma).cpu().numpy()
 
         # Take bottom k% as the attack score
         num_k = max(1, int(len(scores) * self.k))
